Added LambdaEdge to CloudFront for handling URLs that do not end in index.html
