+++ docker-stop.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 
 docker-compose stop
+docker-compose rm -f
