+++ test/json/cf_certificate/update_certificate_change.json
@@ -1,0 +1,24 @@
+{
+    "RequestType": "Update",
+    "ServiceToken": "arn:aws:lambda:us-east-2:1111111111111:function:parima-CertificateLambda-1SWLKCV7LIBSE",
+    "ResponseURL": "http://localhost:9001/responseURL",
+    "StackId": "arn:aws:cloudformation:us-east-2:1111111111111:stack/parima/2d7085e0-13a6-11eb-8119-06104a04a202",
+    "RequestId": "668d8b1b-eb7e-4b1f-acd6-1a02532a76da",
+    "LogicalResourceId": "Certificate",
+    "PhysicalResourceId": "arn:aws:cloudformation:us-east-2:1111111111111:stack/parima/2d7085e0-13a6-11eb-8119-06104a04a202-Certificate",
+    "ResourceType": "Custom::Certificate",
+    "ResourceProperties": {
+        "ServiceToken": "arn:aws:lambda:us-east-2:1111111111111:function:parima-CertificateLambda-1SWLKCV7LIBSE",
+        "OutputParameter": "Certificate",
+        "DomainName": "useast3.tryformkiq.com",
+        "Region": "us-east-1",
+        "StackName": "parima"
+    },
+    "OldResourceProperties": {
+        "ServiceToken": "arn:aws:lambda:us-east-2:1111111111111:function:parima-CertificateLambda-1SWLKCV7LIBSE",
+        "OutputParameter": "Certificate",
+        "DomainName": "useast2.tryformkiq.com",
+        "Region": "us-east-1",
+        "StackName": "parima"
+    }
+}
