+++ test/json/cf_lambda/update.json
@@ -1,0 +1,27 @@
+{
+    "RequestType": "Update",
+    "ServiceToken": "arn:aws:lambda:us-east-2:622653865277:function:parima-LambdaInstall-12GZ8C5I7I95B",
+    "ResponseURL": "http://localhost:9001/responseURL",
+    "StackId": "arn:aws:cloudformation:us-east-2:622653865277:stack/parima/aecebea0-1594-11eb-a6e1-0611fd7a7fc8",
+    "RequestId": "52e3173d-ce27-4bc6-bb82-6a00005ceece",
+    "LogicalResourceId": "CloudFrontRedirectLambda",
+    "ResourceType": "Custom::CreateLambda",
+    "ResourceProperties": {
+        "ServiceToken": "arn:aws:lambda:us-east-2:622653865277:function:parima-LambdaInstall-12GZ8C5I7I95B",
+        "Role": "arn:aws:iam::622653865277:role/parima-CloudFrontRedirectLambdaRole-1XK848KMX6I3Z",
+        "Runtime": "nodejs12.x",
+        "Timeout": "300",
+        "Region": "us-east-1",
+        "MemorySize":"256",
+        "Handler": "index.handler",
+        "Code": {
+            "ZipFile": "UEsDBBQAAAgIAH12XFGOTbdzhwAAALIAAAAIAAAAaW5kZXguanNVykEKwjAQAMCvlCAloWXr2RL9g9dSoSQrCbRJ3d1ARfy7Igh6HUYVxoqFohPVL9mXGQG3NZMwhCn5GclOfE9OU4vGHh8uJ5ZKLMEZXSbPw34EdwXCW0GWnlAKpUqgUIRlEhe0gqbbKXP6WGNVTB43CLLM6vD3uuECY/Ouda2/ufvZppVn/wJQSwECPwMUAAAICAB9dlxRjk23c4cAAACyAAAACAAAAAAAAAAAAAAApIEAAAAAaW5kZXguanNQSwUGAAAAAAEAAQA2AAAArQAAAAAA"
+        },
+        "Tags": [
+            {
+                "Value": "Parima",
+                "Key": "Application"
+            }
+        ]
+    }
+}
